refactor(parsers): extract skill metadata checks into skill_audit_metadata
Move _check_metadata_quality and its _RUNTIME_DEP_PATTERNS /
_REQUIRED_CAPABILITY_KEYS constants out of skill_audit.py into a focused
skill_audit_metadata module. Pure mechanical move, zero behavior change:
audit_skill_result imports the metadata check from the new module.

skill_audit.py drops from 683 to 452 LOC (under the <500 monolith bar);
the metadata quality checks are now independently testable.

Part of #1522
